Maybe fix #20 (WinUSB)
https://social.msdn.microsoft.com/Forums/sqlserver/en-US/ae64282c-3bc3-49af-8391-4d174479d9e7/microsoft-os-20-descriptors-not-working-on-an-interface-of-a-composite-usb-device?forum=wdk

QUOTE
It looks like we'll need to update the MSOS 2.0 Descriptor docs to match
the implementation in USBCCGP.  The bConfigurationValue in the
configuration subset header should actually just be an index value, not
the configuration value.

Specifically it's the index value passed to GET_DESCRIPTOR to retrieve
the configuration descriptor.

Try changing the value to 0 and see if that resolves the issue.  Sorry
for the confusion.
/QUOTE