Use device serialNumber as dictionary key.
In 9b22f8a0e0ee45612cea9dfeecd1b0f1167b15d1 when I was preparing for the
removal of the guid attribute from USBDevice I assumed that the
USBDevice object itself could be used as a functional Object key. This
is not the case (all objects stringify the to same value). So instead we
should be using the device serial number since all WebLights have a
valid and unique serial number.

This resolves issue #41.